[dev] Resume API ReadPage 기능 완성

+++ src/main/java/com/carero/dto/ResultPaging.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class ResultPaging<T> {
-    private int totalCount;
+    private long totalCount;
     private int page;
     private T data;
 }
